working_requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
aistudio-sdk==0.3.8
albucore==0.0.24
albumentations==2.0.8
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.11.0
appdirs==1.4.4
asgiref==3.11.0
astor==0.8.1
attrs==25.4.0
authlib==1.6.5
backoff==2.2.1
bce-python-sdk==0.9.53
bcrypt==5.0.0
beautifulsoup4==4.14.2
blake3==1.0.8
blis==1.3.3
build==1.3.0
cachetools==6.2.2
catalogue==2.0.10
cbor2==5.7.1
certifi==2025.11.12
cffi==2.0.0
cfgv==3.5.0
chardet==5.2.0
charset-normalizer==3.4.4
chromadb==1.1.1
click==8.3.1
cloudpathlib==0.23.0
cloudpickle==3.1.2
coloredlogs==15.0.1
colorlog==6.10.1
compressed-tensors==0.11.0
confection==0.1.5
contourpy==1.3.3
crewai==1.3.0
cryptography==46.0.3
cupy-cuda12x==13.6.0
cycler==0.12.1
cymem==2.0.13
dataclasses-json==0.6.7
deep-translator==1.11.4
deprecation==2.1.0
depyf==0.19.0
dill==0.4.0
diskcache==5.6.3
distlib==0.4.0
distro==1.9.0
dnspython==2.8.0
docstring-parser==0.17.0
doctr==1.9.0
docx2pdf==0.1.8
durationpy==0.10
easyocr==1.7.2
einops==0.8.1
email-validator==2.3.0
et-xmlfile==2.0.0
faiss-cpu==1.12.0
fastapi==0.120.4
fastapi-cli==0.0.16
fastapi-cloud-cli==0.5.1
fastar==0.7.0
fastrlock==0.8.3
fastuuid==0.14.0
filelock==3.20.0
flatbuffers==25.9.23
fonttools==4.60.1
frozendict==2.4.7
frozenlist==1.8.0
fsspec==2025.10.0
future==1.0.0
gguf==0.17.1
google-auth==2.43.0
googleapis-common-protos==1.72.0
greenlet==3.2.4
grpcio==1.76.0
h11==0.16.0
hf-transfer==0.1.9
hf-xet==1.2.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.36.0
humanfriendly==10.0
identify==2.6.15
idna==3.11
imageio==2.37.0
imagesize==1.4.1
importlib-metadata==8.7.0
importlib-resources==6.5.2
instructor==1.13.0
interegular==0.3.3
jinja2==3.1.6
jiter==0.11.1
joblib==1.5.2
json-repair==0.25.2
json5==0.12.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kafka-python==2.2.15
kiwisolver==1.4.9
kornia==0.8.1
kornia-rs==0.1.10
kubernetes==33.1.0
langchain==1.0.3
langchain-classic==1.0.0
langchain-community==0.4.1
langchain-core==1.1.0
langchain-huggingface==1.0.0
langchain-openai==1.0.1
langchain-text-splitters==1.0.0
langcodes==3.5.0
langdetect==1.0.9
langgraph==1.0.2
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.2.10
langsmith==0.4.47
language-data==1.3.0
lark==1.2.2
lazy-loader==0.4
litellm==1.79.1
llguidance==0.7.30
llvmlite==0.44.0
lm-format-enforcer==0.11.3
loguru==0.7.3
lxml==6.0.2
marisa-trie==1.3.1
markdown-it-py==4.0.0
markupsafe==3.0.3
marshmallow==3.26.1
matplotlib==3.10.7
mcp==1.22.0
mdurl==0.1.2
mistral-common==1.8.5
mmh3==5.2.0
modelscope==1.32.0
mpmath==1.3.0
msgpack==1.1.2
msgspec==0.20.0
multidict==6.7.0
murmurhash==1.0.15
mypy-extensions==1.1.0
mysql-connector-python==9.5.0
naked==0.1.32
narwhals==2.12.0
networkx==3.6
ninja==1.13.0
nltk==3.9.2
nodeenv==1.9.1
numba==0.61.2
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
oauthlib==3.3.1
omegaconf==2.3.0
onnxruntime==1.23.2
openai==2.8.1
openai-harmony==0.0.8
opencv-contrib-python==4.10.0.84
opencv-python-headless==4.12.0.88
openpyxl==3.1.5
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-grpc==1.38.0
opentelemetry-exporter-otlp-proto-http==1.38.0
opentelemetry-instrumentation==0.59b0
opentelemetry-instrumentation-asgi==0.59b0
opentelemetry-instrumentation-fastapi==0.59b0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
opentelemetry-util-http==0.59b0
orjson==3.11.4
ormsgpack==1.12.0
outlines-core==0.2.11
overrides==7.7.0
packaging==25.0
paddleocr==3.3.1
paddlex==3.3.10
pandas==2.3.3
partial-json-parser==0.2.1.1.post7
pdf2image==1.17.0
pdfminer-six==20250506
pdfplumber==0.11.7
pillow==12.0.0
pip==24.3.1
platformdirs==4.5.0
plotly==6.3.1
portalocker==2.7.0
posthog==5.4.0
pre-commit==4.5.0
preshed==3.0.12
prettytable==3.17.0
prometheus-client==0.23.1
prometheus-fastapi-instrumentator==7.1.0
propcache==0.4.1
protobuf==6.33.1
psutil==7.1.2
py-cpuinfo==9.0.0
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1-modules==0.4.2
pybase64==1.4.2
pyclipper==1.3.0.post6
pycountry==24.6.1
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.12.3
pydantic-core==2.41.4
pydantic-extra-types==2.10.6
pydantic-settings==2.11.0
pygments==2.19.2
pyjwt==2.10.1
pymupdf==1.26.5
pymysql==1.1.2
pyparsing==3.2.5
pypdf2==3.0.1
pypdfium2==5.0.0
pypika==0.48.9
pyproject-hooks==1.2.0
pytesseract==0.3.13
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
python-json-logger==4.0.0
python-multipart==0.0.20
pytz==2025.2
pyyaml==6.0.2
pyzmq==27.1.0
ray==2.51.1
referencing==0.37.0
regex==2025.11.3
reportlab==4.4.5
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.2.0
rich-toolkit==0.16.0
rignore==0.7.6
rpds-py==0.29.0
rsa==4.9.1
ruamel-yaml==0.18.16
ruamel-yaml-clib==0.2.15
safetensors==0.6.2
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.16.3
seaborn==0.13.2
sentence-transformers==5.1.2
sentencepiece==0.2.1
sentry-sdk==2.43.0
setproctitle==1.3.7
setuptools==79.0.1
shapely==2.1.2
shellescape==3.8.1
shellingham==1.5.4
simsimd==6.5.3
six==1.17.0
smart-open==7.5.0
sniffio==1.3.1
soundfile==0.13.1
soupsieve==2.8
soxr==1.0.0
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlalchemy==2.0.44
srsly==2.5.2
sse-starlette==3.0.3
starlette==0.49.3
stringzilla==4.2.3
sympy==1.14.0
tenacity==9.1.2
textblob==0.19.0
thinc==8.3.10
threadpoolctl==3.6.0
tifffile==2025.10.16
tiktoken==0.12.0
tokenizers==0.22.1
tomli==2.3.0
tomli-w==1.2.0
torch==2.8.0
torchaudio==2.8.0
torchvision==0.23.0
tqdm==4.67.1
transformers==4.57.1
triton==3.4.0
ty==0.0.1a27
typer==0.20.0
typer-slim==0.20.0
typing-extensions==4.15.0
typing-inspect==0.9.0
typing-inspection==0.4.2
tzdata==2025.2
ujson==5.11.0
urllib3==2.5.0
uv==0.9.11
uvicorn==0.38.0
uvloop==0.22.1
validators==0.35.0
virtualenv==20.35.4
vllm==0.11.0
wasabi==1.1.3
watchfiles==1.1.1
wcwidth==0.2.14
weasel==0.4.3
weaviate-client==4.17.0
websocket-client==1.9.0
websockets==15.0.1
wrapt==1.17.3
xformers==0.0.32.post1
xgrammar==0.1.25
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0
